+++ tests/ert/unit_tests/gui/ertwidgets/test_storage_info_widget.py
@@ -14,8 +14,8 @@ from ert.ensemble_evaluator import state
 from ert.gui.experiments.view import run_status as run_status_module
 from ert.gui.experiments.view.realization import RealizationWidget
 from ert.gui.experiments.view.run_status import RunStatusView
-from ert.gui.tools.manage_experiments.storage_info_widget import (
-    _EnsembleWidget,
+from ert.gui.tools.manage_experiments.ensemble_widget import (
+    EnsembleWidget,
     _EnsembleWidgetTabs,
 )
 from ert.run_models.event import FullSnapshotEvent, status_event_to_json
@@ -87,7 +87,7 @@ def test_that_missing_response_for_observation_response_key_does_not_crash(
         0,
     )
 
-    ensemble_widget = _EnsembleWidget()
+    ensemble_widget = EnsembleWidget()
     ensemble_widget.setEnsemble(ensemble)
     qtbot.addWidget(ensemble_widget)
 
@@ -134,7 +134,7 @@ def test_that_breakthrough_experiment_does_not_crash(qtbot, storage):
     breakthrough_response = bt_config.derive_from_storage(0, 0, ensemble)
     ensemble.save_response("breakthrough", breakthrough_response, 0)
 
-    ensemble_widget = _EnsembleWidget()
+    ensemble_widget = EnsembleWidget()
     ensemble_widget.setEnsemble(ensemble)
     qtbot.addWidget(ensemble_widget)
 
@@ -219,7 +219,7 @@ def test_that_rft_experiment_without_a_zone_does_not_crash(qtbot, storage):
     ensemble.save_response("rft", rft_response(), 0)
     ensemble.save_observation_location_metadata(location_metadata(), 0)
 
-    ensemble_widget = _EnsembleWidget()
+    ensemble_widget = EnsembleWidget()
     ensemble_widget.setEnsemble(ensemble)
     qtbot.addWidget(ensemble_widget)
 
@@ -312,7 +312,7 @@ def test_that_approximated_rft_responses_are_visualized_in_ensemble_widget_if_en
     ensemble.save_response("rft", rft_responses, 0)
     ensemble.save_observation_location_metadata(location_metadata, 0)
 
-    ensemble_widget = _EnsembleWidget()
+    ensemble_widget = EnsembleWidget()
     ensemble_widget.setEnsemble(ensemble)
     qtbot.addWidget(ensemble_widget)
 
@@ -461,7 +461,7 @@ def test_that_many_realizations_in_rft_affect_responses_not_observation_tree(
         location_metadata(actual_zones=zones2, well_connection_cell=cell2), 3
     )
 
-    ensemble_widget = _EnsembleWidget()
+    ensemble_widget = EnsembleWidget()
     ensemble_widget.setEnsemble(ensemble)
     qtbot.addWidget(ensemble_widget)
 
@@ -519,7 +519,7 @@ def test_that_both_observations_with_same_data_are_displayed(qtbot, storage):
     experiment = create_experiment_from_config(config, storage)
     ensemble = experiment.create_ensemble(name="default", ensemble_size=1)
 
-    ensemble_widget = _EnsembleWidget()
+    ensemble_widget = EnsembleWidget()
     ensemble_widget.setEnsemble(ensemble)
     qtbot.addWidget(ensemble_widget)
 
@@ -674,7 +674,7 @@ def test_that_observations_are_identified_and_sorted_by_full_index_key(
     experiment = create_experiment_from_config(config, storage)
     ensemble = experiment.create_ensemble(name="default", ensemble_size=1)
 
-    ensemble_widget = _EnsembleWidget()
+    ensemble_widget = EnsembleWidget()
     ensemble_widget.setEnsemble(ensemble)
     qtbot.addWidget(ensemble_widget)
 
@@ -728,7 +728,7 @@ def test_that_run_status_tab_shows_persisted_realizations_for_ensemble(qtbot, st
     ensemble = experiment.create_ensemble(name="default", ensemble_size=2, iteration=0)
     _persist_full_snapshot(experiment, ensemble.iteration)
 
-    ensemble_widget = _EnsembleWidget()
+    ensemble_widget = EnsembleWidget()
     ensemble_widget.setEnsemble(ensemble)
     qtbot.addWidget(ensemble_widget)
 
@@ -751,7 +751,7 @@ def test_that_run_status_tab_shows_placeholder_when_no_snapshot_exists(qtbot, st
     experiment = storage.create_experiment(name="exp")
     ensemble = experiment.create_ensemble(name="default", ensemble_size=2, iteration=0)
 
-    ensemble_widget = _EnsembleWidget()
+    ensemble_widget = EnsembleWidget()
     ensemble_widget.setEnsemble(ensemble)
     qtbot.addWidget(ensemble_widget)
